Removido classe Scheduler
-- Removido a classe Scheduler, pois ela foi re-implementada como uma classe Queue
-- Renomeado arquivo "queue", que tinha as classes de Queue, para scheduler, pois essas classes fazem o agendamento de transações para o withdraw queue, e tbm pq "scheduler" é um nome muito mais legal
