+++ external_modules/common/index.ts
@@ -4,12 +4,12 @@ import io from 'socket.io-client'
 import { EventEmitter } from 'events'
 import { startSession } from 'mongoose'
 import Sync from './sync'
-import { Single } from './queue'
+import { Single } from './scheduler'
 import initListeners from './listeners'
 import Transaction, { CreateSendRequest, Receive, Send, SendRequestDoc } from './db/models/transaction'
 import type TypedEmitter from 'typed-emitter'
 import type { ClientSession } from 'mongoose'
-import type { Queue } from './queue'
+import type { Queue } from './scheduler'
 import type { WithdrawRequest as WR } from '../../interfaces/transaction'
 import type { ExternalEvents } from '../../interfaces/communication/external-socket'
 import type { ReceiveDoc, SendDoc, CreateReceive, CreateSend } from './db/models/transaction'
